currency/requirements.txt
Django==3.0.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
pprint==0.1
requests==2.23.0
gunicorn==20.0.4
